GH-40 Fix encoding/decoding variable length field (#43)
The packet of every header includes a field "variable length".
This field encodes the number of bytes that follow.

The field's size is variable between 1 and including 4 bytes,
depending on the length of the packet.

The encoded and decoder for the variable length field incorrectly
handled this field when it was 4 bytes long.

Among others, that prevented users from building
a `Subscribe` packet with a large number of topics.