GH-55 Fix hot loop when keep alive interval is 0
MQTT implements a mechanism to keep the connection active.
Every X seconds, a message must be emitted by the client.
The interval is defined by the keep alive setting.

When this setting is configured to 0 seconds, this mechanism is deactivated:

> A Keep Alive value of zero (0) has the effect of turning off the keep alive mechanism.

Until now, the MqttBinding didn't implement this requirement.
Instead, it goes in a hot loop emitting PINGREQ messages.

That has been fixed by this commit.