GH-74 `Client` to support QoS 1 and QoS 2 (#75)
Whn a peer emits a PUBLISH packet with QOS 1,
the receiving peer must acknowledge the message
with a PUBACK.

When a peer emits a PUBLISH packe with QOS 2,
there is a sequence of PUBREC, PUBREL and PUBCOMP.

This commit adds support for these messages to the `Client`.

This commit also adds the test helper "wiretapped_client".
This function creates a regular `Client`. All `Packet`s
from and to this `Client` are stored in ` Transcription`.

The `Transcription` records all messages sent between
a MQTT client and the broker.

This type can be used to introspect that traffic. In particular
to look for messages that are not "leaking" to the `Client`.

For example, upon receiving a PUBLISH with  QoS 1 or 2,
the `Client` must emit respectively a PUBACK or PUBREC.

There is no way to verify these messages have been sent, other
than via the `Transcription`.