GH-84 Remove `impl Clone` from `aio::ClientHandler` (#90)
As result, handler's inner broadcast channel can be replaced with a regular `Receiver`. That affects
the error type returned by `ClientHandler::publication()`. The error type has been changed from `broadcast_channel::RecvError` to `HandlerError`. That is the same error as returned by `blocking::ClientHandler::publication()`.

I removed `Clone` because it cloning the handler
introduces some surprises.

For example, consider you've 2 handlers `a` and `b`. `a` subscribes to "topic-1" and `b` subscribes to
topic "topic-2". Messages to any topic all distributed to all handlers. So `a` recieves messages for "topic-b". That surprising, at least.

Slow handlers can cause blockages for all handlers. The `broadcast` channel has a fixed capacity. The `Client` can only push new packets to the channel if there is space left. If one handler is slow to consume packets and the channel is full, none of the other handlers receive new packets.

Lastly, removing `Clone` removes the need for the dependency `async_broadcast`.