GH-100 Add an experimental server
This commit adds a broker. At the moment the broker
supports:

* CONNECT without any flags set
* SUBSCRIBE with QoS 0
* PUBLISH with QoS 0

Run the broker with:

```
$ cargo run --features="async experimental" --example server
```

And connect a client:

```
$ cargo run --example client_with_smol localhost:1883
```

The server is in a very early stage. Most of the
MQTT specification is not implemented. And it crashes
easily. For example, using `mosquitto_pub` crashes
the server.

I'm _not_ planning to implement the complete
 MQTT spec. For now,  I aim to support
a handfull of clients using QoS 0 exclusively.