GH-3 Add support for DISCONNECT
Currently, the crate doesn't implement some reconnecting strategy.
If a TCP connection dies, `Client.run()` stops with an error and
the `ClientHandler` stops working.

One need to recreate a `Client` and call `Client::spawn()` to obtain
a `ClientHandler`.

In other words, it is the responsibility of the user to handle a
failing connection.

I think I'm fine with that for now.

The `Client`s are generic over the Socket `S`. `S` must implement `Read`
and `Write` or their async counter parts. So `Client` can't reconnect.

If there was a common trait that implement reconnecting, we could expand
the trait bounds of `S` to `Read + Write + Reconnect`. But I'm not aware
of such trait in the async or blocking ecosystem.